Push proof dependencies from SignNewDocumentsTask to proof package
